ms/vehicle_ms/requirements.txt
Flask==3.0.2
flask-marshmallow==0.15.0
marshmallow==3.20.1
werkzeug>=3.0.1 # not directly required, pinned by Snyk to avoid a vulnerability
python-dotenv == 0.19.0
proofdock-chaos-middleware-flask==0.1.3
proofdock-chaos-middleware-python==0.2.3
flask_sqlalchemy == 3.0.5
sqlalchemy_filters == 0.13.0
flask-migrate == 4.0.4
psycopg2 == 2.9.7
